Add namespace to all yaml files.
There was one yaml that had namespace hardocded to "default" and the
rest had nothing. A bug was reported where "project" was set to
something other than "default", so all objects were created in the
"project" namespace except the one hardcoded. Made namespace a variable
so it can be overwritten, but defaults to "default".

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>